+++ main.py
@@ -15,9 +15,9 @@ import random
 from Score import save_score, show_leaderboard
 
 from Shelf import Shelf
-# from Body import Body
 from Instruction import show_instructions
 from Useful import get_player_name
+from Window import DrawWindow, ScreenRollDown
 
 
 def show_difficulty_selection():
@@ -171,51 +171,6 @@ def HandleMovement(keys_pressed):  # Handling the Left/Right buttons pressing.
             body.acceleration = MAX_ACCELERATION
 
 
-def DrawWindow():  # Basically, drawing the screen.
-    global WALLS_Y
-    font = pygame.font.SysFont("Arial", 26)
-    HandleBackground()
-    for shelf in total_shelves_list:
-        if shelf.number < 20:
-            shelf_image = SHELF_BRICK_IMAGE
-        elif shelf.number < 40:
-            shelf_image = SHELF_BRICK_IMAGE2
-        else:
-            shelf_image = SHELF_BRICK_IMAGE3
-
-        for x in range(shelf.rect.x, shelf.rect.x + shelf.width, 32):
-            WIN.blit(shelf_image, (x, shelf.rect.y))  # Drawing the shelf.
-            if shelf.number % 10 == 0 and shelf.number != 0:
-                shelf_number = pygame.Rect(shelf.rect.x + shelf.rect.width / 2 - 16, shelf.rect.y,
-                                           16 * len(str(shelf.number)), 25)
-                pygame.draw.rect(WIN, GRAY, shelf_number)
-                txt = font.render(str(shelf.number), True, BLACK)
-                WIN.blit(txt,
-                         (shelf.rect.x + shelf.rect.width / 2 - 16, shelf.rect.y))  # Drawing the number of the shelf.
-    for y in range(WALLS_Y, HEIGHT, 108):  # Drawing the walls.
-        WIN.blit(BRICK_IMAGE, (0, y))
-        WIN.blit(BRICK_IMAGE, (WIDTH - WALL_WIDTH, y))
-
-    rotated_image = pygame.transform.rotate(BODY_IMAGE, body.angle)
-    new_rect = rotated_image.get_rect(center=BODY_IMAGE.get_rect(topleft=(body.x, body.y)).center)
-
-    WIN.blit(rotated_image, new_rect.topleft)  # Draw the rotated character
-    pygame.display.update()
-
-
-def ScreenRollDown():  # Increasing the y values of all elements.
-    global background_y, WALLS_Y
-    for shelf in total_shelves_list:
-        shelf.rect.y += 1
-    body.y += 1
-    background_y += 0.5
-    if background_y == BACKGROUND_Y + 164:
-        background_y = BACKGROUND_Y
-    WALLS_Y += WALLS_ROLLING_SPEED
-    if WALLS_Y == 0:
-        WALLS_Y = -108
-
-
 def GameOver():
     print("Game Over")
     print("Your score is: ", MAX_SHELF_NUMBER)
@@ -238,11 +193,6 @@ def CheckIfTouchingFloor():  # Checking if the body is still on the main ground.
             standing, falling = True, False
         else:  # In a more advanced part of the game, when can already lose.
             GameOver()
-
-
-def HandleBackground():  # Drawing the background.
-    if body.y >= total_shelves_list[500].rect.y:
-        WIN.blit(BACKGROUND, (32, background_y))
 
 
 def OnShelf():  # Checking whether the body is on a shelf, returning True/False.
